Admin permissions

+++ app/Containers/Role/UI/API/Controllers/RoleController.php
@@ -4,19 +4,22 @@ namespace App\Containers\Role\UI\API\Controllers;
 
 use App\Containers\Role\Models\Role;
 use App\Containers\Role\UI\API\Requests\RoleRequest;
+use App\Containers\Role\UI\API\Resources\RoleResource;
 use App\Ship\Parents\Controllers\ApiController;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
-use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends ApiController
 {
-
-    /**
-     * @return Collection
-     */
-    public function index(): Collection
+     /**
+     * @return AnonymousResourceCollection
+      */
+    public function index(): AnonymousResourceCollection
     {
-        return Role::all();
+        Gate::authorize('view', 'roles');
+
+        return RoleResource::collection(Role::all());
     }
 
     /**
@@ -25,17 +28,31 @@ class RoleController extends ApiController
      */
     public function store(RoleRequest $request): Response
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
-        return response($role, Response::HTTP_CREATED);
+
+        if ($permissions = $request->input('permissions')) {
+            foreach ($permissions as $permission_id) {
+                \DB::table('role_permission')->insert([
+                   'role_id' => $role->id,
+                   'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     /**
      * @param int $id
-     * @return Role
+     * @return RoleResource
      */
-    public function show(int $id): Role
+    public function show(int $id): RoleResource
     {
-        return Role::find($id);
+        Gate::authorize('view', 'roles');
+
+        return new RoleResource(Role::find($id));
     }
 
     /**
@@ -45,10 +62,25 @@ class RoleController extends ApiController
      */
     public function update(RoleRequest $request, int $id): Response
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
         $role->update($request->only('name'));
 
-        return response($role, Response::HTTP_ACCEPTED);
+        \DB::table('role_permissions')
+            ->where('role_id', $role->id)
+            ->delete();
+
+        if ($permissions = $request->input('permissions')) {
+            foreach ($permissions as $permission_id) {
+                \DB::table('role_permission')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
     /**
@@ -57,6 +89,12 @@ class RoleController extends ApiController
      */
     public function destroy(int $id): Response
     {
+        Gate::authorize('edit', 'roles');
+
+        \DB::table('role_permissions')
+            ->where('role_id', $id)
+            ->delete();
+
         Role::destroy($id);
         return response(null, Response::HTTP_NO_CONTENT);
     }
